+++ ScriperSol/Scriper/ViewModels/ScriptManagerVM.cs
@@ -5,8 +5,8 @@ using NLog;
 using ReactiveUI;
 using Scriper.AssetsAccess;
 using Scriper.Converters;
+using Scriper.CustomScripts;
 using Scriper.Extensions;
-using Scriper.Models;
 using Scriper.SystemTray;
 using Scriper.TimeSchedule;
 using Scriper.Views;
